Refactored newWallet
