+++ packages/keymaster/src/db-wallet-web.js
@@ -1,11 +1,12 @@
-import * as cipher from '@macterra/cipher/web';
-import * as exceptions from '@macterra/exceptions';
-
 const walletName = 'mdip-keymaster';
 
-export function saveWallet(wallet) {
-    // TBD validate wallet before saving
+export function saveWallet(wallet, overwrite = false) {
+    if (!overwrite && window.localStorage.getItem(walletName)) {
+        return false;
+    }
+
     window.localStorage.setItem(walletName, JSON.stringify(wallet));
+    return true;
 }
 
 export function loadWallet() {
@@ -16,34 +17,4 @@ export function loadWallet() {
     }
 
     return newWallet();
-}
-
-export function newWallet(mnemonic, overwrite = false) {
-    if (!overwrite && window.localStorage.getItem(walletName)) {
-        throw new Error(exceptions.UPDATE_FAILED);
-    }
-
-    try {
-        if (!mnemonic) {
-            mnemonic = cipher.generateMnemonic();
-        }
-        const hdkey = cipher.generateHDKey(mnemonic);
-        const keypair = cipher.generateJwk(hdkey.privateKey);
-        const backup = cipher.encryptMessage(keypair.publicJwk, keypair.privateJwk, mnemonic);
-
-        const wallet = {
-            seed: {
-                mnemonic: backup,
-                hdkey: hdkey.toJSON(),
-            },
-            counter: 0,
-            ids: {},
-        }
-
-        saveWallet(wallet);
-        return wallet;
-    }
-    catch (error) {
-        throw new Error(exceptions.INVALID_PARAMETER);
-    }
 }
